Rename IndexConfig to IndexConfigTrait
And restore CoveringIndexConfig to IndexConfig, to keep the user
interface compatibility
