+++ python/hyperspace/tests/test_indexutilization.py
@@ -40,7 +40,7 @@ class HyperspaceIndexUtilizationTests(HyperspaceTestCase):
                          "Hyperspace must be disabled after Hyperspace disable.")
 
     def test_hyperspace_explain(self):
-        idx_config = CoveringIndexConfig('idx1', ['age'], ['name', 'department'])
+        idx_config = IndexConfig('idx1', ['age'], ['name', 'department'])
         self.hyperspace.createIndex(self.df, idx_config)
         self.df.createOrReplaceTempView("employees")
         filter_query = self.spark.sql(""" SELECT age, name, department FROM employees
