+++ src/main/scala/com/microsoft/hyperspace/Hyperspace.scala
@@ -40,7 +40,7 @@ class Hyperspace(spark: SparkSession) {
    * @param df the DataFrame object to build index on.
    * @param indexConfig the configuration of index to be created.
    */
-  def createIndex(df: DataFrame, indexConfig: IndexConfig): Unit = {
+  def createIndex(df: DataFrame, indexConfig: IndexConfigTrait): Unit = {
     withHyperspaceRuleDisabled {
       indexManager.create(df, indexConfig)
     }
