+++ src/test/scala/com/microsoft/hyperspace/index/HybridScanForDeltaLakeTest.scala
@@ -44,7 +44,7 @@ class HybridScanForDeltaLakeTest extends HybridScanSuite {
   override def setupIndexAndChangeData(
       sourceFileFormat: String,
       sourcePath: String,
-      indexConfig: IndexConfig,
+      indexConfig: IndexConfigTrait,
       appendCnt: Int,
       deleteCnt: Int): (Seq[String], Seq[String]) = {
     // Use partition to control the number of appended and deleted files.
