+++ src/test/scala/com/microsoft/hyperspace/index/IcebergIntegrationTest.scala
@@ -35,8 +35,6 @@ import com.microsoft.hyperspace.util.PathUtils
 import com.microsoft.hyperspace.util.PathUtils.DataPathFilter
 
 class IcebergIntegrationTest extends QueryTest with HyperspaceSuite {
-  private val IndexConfig = CoveringIndexConfig
-
   override val indexLocationDirName = "icebergIntegrationTest"
 
   private val sampleData = SampleData.testData
