+++ src/test/scala/com/microsoft/hyperspace/index/IndexCacheTest.scala
@@ -21,13 +21,13 @@ import org.apache.hadoop.yarn.util.Clock
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.types.{StringType, StructField, StructType}
 
-import com.microsoft.hyperspace.{Hyperspace, HyperspaceException, SampleData, TestCoveringIndex}
+import com.microsoft.hyperspace.{Hyperspace, HyperspaceException, SampleData}
 import com.microsoft.hyperspace.actions.Constants
 import com.microsoft.hyperspace.util.FileUtils
 
 class IndexCacheTest extends HyperspaceSuite {
   val sampleParquetDataLocation = inTempDir("sampleparquet")
-  val indexConfig1 = CoveringIndexConfig("index1", Seq("RGUID"), Seq("Date"))
+  val indexConfig1 = IndexConfig("index1", Seq("RGUID"), Seq("Date"))
 
   before {
     FileUtils.delete(new Path(sampleParquetDataLocation))
@@ -48,11 +48,12 @@ class IndexCacheTest extends HyperspaceSuite {
 
     val entry = IndexLogEntry(
       "index1",
-      TestCoveringIndex(
+      CoveringIndex(
         Seq("RGUID"),
         Seq("Date"),
         schema,
-        10),
+        10,
+        Map()),
       Content(Directory(indexDir)),
       Source(SparkPlan(sourcePlanProperties)),
       Map())
