+++ src/test/scala/com/microsoft/hyperspace/index/RefreshIndexNestedTest.scala
@@ -35,8 +35,7 @@ class RefreshIndexNestedTest extends QueryTest with HyperspaceSuite {
   private val testDir = inTempDir("RefreshIndexDeleteTests")
   private val nonPartitionedDataPath = testDir + "/nonpartitioned"
   private val partitionedDataPath = testDir + "/partitioned"
-  private val indexConfig =
-    CoveringIndexConfig("index1", Seq("nested.leaf.id"), Seq("nested.leaf.cnt"))
+  private val indexConfig = IndexConfig("index1", Seq("nested.leaf.id"), Seq("nested.leaf.cnt"))
   private var hyperspace: Hyperspace = _
 
   override def beforeAll(): Unit = {
