+++ src/test/scala/com/microsoft/hyperspace/index/rules/CandidateIndexCollectorTest.scala
@@ -26,12 +26,10 @@ import org.apache.spark.sql.types.{IntegerType, StringType}
 import com.microsoft.hyperspace.Hyperspace
 import com.microsoft.hyperspace.TestUtils.latestIndexLogEntry
 import com.microsoft.hyperspace.actions.Constants
-import com.microsoft.hyperspace.index.{CoveringIndexConfig, IndexCollectionManager, IndexConstants, IndexLogEntryTags}
+import com.microsoft.hyperspace.index.{IndexCollectionManager, IndexConfig, IndexConstants, IndexLogEntryTags}
 import com.microsoft.hyperspace.util.FileUtils
 
 class CandidateIndexCollectorTest extends HyperspaceRuleSuite with SQLHelper {
-  private val IndexConfig = CoveringIndexConfig
-
   override val indexLocationDirName = "candidateIndexCollectorTest"
 
   val t1c1 = AttributeReference("t1c1", IntegerType)()
